json manager and clean architecture
