+++ app/src/main/java/com/example/dispatcher/data/FirebaseCrashlyticsManager.kt
@@ -1,7 +1,6 @@
-package com.example.dispatcher
+package com.example.dispatcher.data
 
 import android.content.Context
-import android.os.Bundle
 import android.util.Log
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.crashlytics.FirebaseCrashlytics
