+++ app/src/main/java/com/example/dispatcher/presentation/SplashActivity.kt
@@ -1,19 +1,10 @@
-package com.example.dispatcher
+package com.example.dispatcher.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.example.dispatcher.ui.theme.DispatcherTheme
 import android.content.Intent
 import android.os.Handler
+import com.example.dispatcher.R
 
 class SplashActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
