+++ app/src/main/java/com/example/dispatcher/presentation/utils/ViewExtensions.kt
@@ -1,4 +1,4 @@
-package com.example.dispatcher
+package com.example.dispatcher.presentation.utils
 
 import android.widget.Toast
 import androidx.fragment.app.Fragment
